updated the run script  run.sh   ANDREA
git-svn-id: svn+ssh://qeforge.qe-forge.org/svnroot/want/trunk@102 0e440500-e291-4b27-a0db-898e6d94e59d

+++ want/tests/test01/run.sh
@@ -87,10 +87,14 @@ esac
 test -e $TMPDIR/$TEST_NAME || mkdir $TMPDIR/$TEST_NAME 
 cd $TMPDIR/$TEST_NAME
 ln -sf $TEST_HOME/../Pseudo/*.UPF .
-cd $TEST_HOME
-ln -sf $TMPDIR/$TEST_NAME ./SCRATCH
-cd $TMPDIR/$TEST_NAME
-ln -sf $TEST_HOME ./HOME
+if [ ! -e $TEST_HOME/SCRATCH ] ; then
+    cd $TEST_HOME
+    ln -sf $TMPDIR/$TEST_NAME ./SCRATCH
+fi
+if [ ! -e $TMPDIR/$TEST_NAME/HOME ] ; then
+    cd $TMPDIR/$TEST_NAME
+    ln -sf $TEST_HOME ./HOME
+fi
 
 
 #-----------------------------------------------------------------------------
